codefactor: move api/mcp/ to clients/ and fix imports.
