+++ eg1/api/endpoints.py
@@ -9,10 +9,10 @@ import eg1.api.crud as crud
 import eg1.api.schemas as schemas
 from eg1.api.db import get_db
 from eg1.api.errors import CustomIntegrityError, SchemaError
-from eg1.api.logger import logger
 from eg1.api.metrics import Metric, metric_registry
-from eg1.runners import dispatch_retries, dispatch_tasks
 from eg1.api.security import admin_only
+from eg1.logger import logger
+from eg1.runners import dispatch_retries, dispatch_tasks
 
 router = APIRouter()
