+++ eg1/clients/__init__.py
@@ -1,3 +1,5 @@
-from .llm import LlmClient
+# <!>WARN<!> Order of import matter to avoid circular import here.
 from .schemas.openai import ChatCompletionRequest, ChatCompletionResponse
 from .schemas.openai_rag import Chunk, RagChatCompletionResponse, Search
+from .llm import LlmClient
+from .mcp import MCPBridgeClient, multi_step_generate
