+++ eg1/clients/mcp.py
@@ -3,8 +3,7 @@ import json
 import requests
 
 from eg1.clients import LlmClient, ChatCompletionResponse
-from eg1.api.logger import logger
-import eg1.api.models as models
+from eg1.logger import logger
 from eg1.utils import log_and_raise_for_status
