+++ eg1/logger.py
