+++ eg1/runners/tasks.py
@@ -10,9 +10,9 @@ import eg1.api.crud as crud
 import eg1.api.models as models
 from eg1.api.config import DEFAULT_JUDGE_MODEL
 from eg1.api.db import SessionLocal
-from eg1.api.logger import logger
 from eg1.api.metrics import metric_registry
-from eg1.mcp import MCPBridgeClient, multi_step_generate
+from eg1.clients import MCPBridgeClient, multi_step_generate
+from eg1.logger import logger
 from eg1.runners import MessageType, dispatch_tasks
 from eg1.utils import Timer, run_with_timeout
